Revert "Don't take ceiling of log2 of registers"
This reverts commit 754438eb16596941a2c85eba56a92d810094a69d.

Not taking the ceiling does the exact opposite of what the commit
message claims. By not taking the ceiling, many values will be
rounded down due to the cast to uint32: uint32(log2(64))=5. This
lead to some m values using only half the requested registers, and
subsequently errors close to 100%.