Fix unsafe pointer usage in MurmurBytes
This commit updates MurmurBytes to read the input data as a block of []uint32 rather than one uint32 at a time.  This allows us to signal to the compiler that the data is aligned in a way that doesn't fail the checkptr check.

This commit also swaps in bits.RotateLeft32 instead of the handrolled version as it is equally fast and there is no need to duplicate stdlib

Profiling revealed that creating the `tail` slice was actually a source of some CPU usage so this commit also removes that and instead calculates the length of the tail and the index where it starts.